Move old heading component to onboarding
