+++ packages/ui/src/index.ts
@@ -14,9 +14,8 @@ export type { InputBaseProps } from './components/InputBase'
 export { InputField } from './components/InputField/InputField'
 export type { InputFieldProps } from './components/InputField/InputField'
 export { InputStepper } from './components/InputStepper/InputStepper'
-export { Heading as HeadingOLD } from './components/Heading/Heading'
-export { Heading } from './components/HeadingNew/HeadingNew'
-export type { HeadingProps } from './components/HeadingNew/HeadingNew'
+export { Heading } from './components/Heading/Heading'
+export type { HeadingProps } from './components/Heading/Heading'
 export * as Dialog from './components/Dialog/Dialog'
 export { HeadingLabel } from './components/HeadingLabel/HeadingLabel'
