+++ packages/ui/src/lib/theme/typography.stories.tsx
@@ -1,4 +1,4 @@
-import { Heading } from '../../components/HeadingNew/HeadingNew'
+import { Heading } from '../../components/Heading/Heading'
 import { theme } from './theme'
 
 const config = {
